+++ src/Events/GuildMemberRemove.ts
@@ -1,17 +1,30 @@
 import { Event } from "../Structures/Event";
-import { ClientEvents, GuildMember } from "discord.js";
+import { ClientEvents, GuildMember, TextChannel, Permissions } from "discord.js";
 import { pool } from "../Structures/Database/Mongo";
 import { formatDate } from "../lib/Utility/Date";
+import { Logger } from "../Structures/Logger";
+import { inspect } from "util";
+import Embed from "../Structures/Embed";
+import { GuildSettings } from "../lib/types/Collections";
+
+const basic = new Permissions([
+    'SEND_MESSAGES',
+    'EMBED_LINKS',
+    'VIEW_CHANNEL'
+]);
 
 export default class implements Event {
     name: keyof ClientEvents = 'guildMemberRemove';
+    logger = new Logger(this.name);
 
     async init(member: GuildMember) {
         const date = formatDate('MM-DD-YYYY', new Date());
         const client = await pool.insights.connect();
-        const collection = client.db('khafrabot').collection('insights');
+        
+        const insightsCollection = client.db('khafrabot').collection('insights');   
+        const settingsCollection = client.db('khafrabot').collection('settings');
             
-        collection.updateOne(
+        insightsCollection.updateOne(
             { id: member.guild.id },
             { $inc: { 
                 [`daily.${date}.total`]: -1,
@@ -19,5 +32,29 @@ export default class implements Event {
             } },
             { upsert: true }
         );
+
+        const server = await settingsCollection.findOne({ id: member.guild.id }) as GuildSettings;
+        if(!server.welcomeChannel) {
+            return;
+        }
+
+        let channel: TextChannel;
+        try {
+            // TextChannel logic is handled where the user sets the channel
+            channel = await member.guild.client.channels.fetch(server.welcomeChannel) as TextChannel;
+        } catch(e) {
+            this.logger.log(inspect(e));
+            return;
+        }
+
+        if(!channel.permissionsFor(member.guild.me).has(basic)) {
+            return;
+        }
+
+        const embed = Embed.success()
+            .setAuthor(member.user.username, member.user.displayAvatarURL())
+            .setDescription(`${member} has left the server!`);
+
+        return channel.send(embed);
     }
 }